test_project/requirements.txt
# django>=1.8, <1.9
# django>=1.7, <1.8
# django>=1.4, <1.5
# django>=1.6, <1.7

# djangorestframework>=2.4, <3.0  # for django<1.8
# djangorestframework>=3.0, <4.0  # for django>=1.8

# dev tools
tox
pytest
mock==1.0.1
six==1.9.0
django-extensions==1.5.5
django-filter==2.1.0
ipython
